Refactor postgresql/user.js model
