+++ src/controllers/user.js
@@ -1,9 +1,6 @@
+import { UserModel } from "../models/postgresql/user.js"
 import { validatePasswordForm, validateUser } from "../schemas/user.js"
-
 import { sendMailVerification } from "../services/mail/sendMailRegister.js"
-
-import { UserModel } from "../models/turso/user.js"
-// import { UserModel } from "../models/postgresql/user.js"
 
 export class UserController {
 	static async init(req, res) {
