+++ src/routes/db.js
@@ -6,6 +6,6 @@ import { UserController } from "../controllers/user.js"
 
 export const dbRouter = Router()
 
-dbRouter.post("/users", UserController.init)
+dbRouter.get("/users", UserController.init)
 dbRouter.post("/subjects", SubjectController.init)
 dbRouter.post("/classes", ClassController.init)
